remove old ksh tty HISTORY and DISPLAY setup
X and DISPLAY aren't really needed anymore and ksh is mostly
a relic of the past.  Having this code here was causing problems
with this .profile being executed on gitpod.
ssh -X  is good enough to set the DISPLAY.  We don't need to
try to automagically create DISPLAY as if we were using telnet
in the past.
I don't think this HISTFILE setting with the tmp dir really
worked well anyway - maybe in limited cases on certain systems
that have long gone away?